Add files via upload
The artists, songs, duration in milliseconds, year the track was released, and popularity of the song are all listed in this column. Dance ability (describe how danceable the track is). (Explain how energy is perceived as a measure of intensity and activity.) the Key, the Loudness in decibels, the Mode (describe the major and minor), the Speechiness (describe the presence of spoken words in the track), the Acoustiness (describe the confidence measure from 0.0 to 1.0), the Liveness (describe the presence of an audience in the recording), the Valence (describe the positiveness conveyed by the track), the Tempo (describe the overall estimate track in beats per minute), and the Genre (describe the Genre of the track